comit provisional 3

+++ includes/clients/LlamaTest.h
@@ -16,7 +16,7 @@ private:
     std::string filepath_; //ruta para guardar los resultados
     long long test_id; // el timestammp de comienczo en la practica
     std::string model_path_;
-    int temperature_;
+    float temperature_;
     int batch_size_;
     int context_size_;
     int seed_;
@@ -24,7 +24,7 @@ private:
     double hardwarePeriod_;
     std::string getCleanModelPath();
 public:
-    LlamaTest(std::string filepath_, std::string model_path, int temperature, int batch_size, int context_size, int seed, int num_prompts, double hardwarePeriod = 0.5);
+    LlamaTest(std::string filepath_, std::string model_path, float temperature, int batch_size, int context_size, int seed, int num_prompts, double hardwarePeriod = 0.5);
     LlamaTest(nlohmann::json);
     bool runTestType0(); // solo procesa los prompts, no toma medidas hardware
     bool runTestType1();//paralelismo an nivel de prompt individual (tras procesar un promp converge los h
